requirements.txt
anthropic==0.39.0
fastapi==0.115.0
httpx==0.27.2
openai==1.54.0
pydantic==2.9.2
pydantic-settings==2.6.1
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
supabase==2.9.1
uvicorn[standard]==0.32.0
